+++ ipfabric/api.py
@@ -50,6 +50,7 @@ class IPFabricAPI(Client):
             snapshot_id: Optional[str] = DEFAULT_ID,
             username: Optional[str] = None,
             password: Optional[str] = None,
+            unloaded: bool = False,
             **kwargs,
     ):
         """
@@ -59,6 +60,7 @@ class IPFabricAPI(Client):
         :param snapshot_id: str: IP Fabric snapshot ID to use by default for database actions - defaults to '$last'
         :param kwargs: dict: Keyword args to pass to httpx
         """
+        self.unloaded = unloaded
         with Settings() as settings:
             super().__init__(
                 timeout=kwargs.get("timeout", True),
@@ -147,6 +149,10 @@ class IPFabricAPI(Client):
 
     @property
     def unloaded_snapshots(self):
+        if not self.unloaded:
+            logger.warning("Unloaded snapshots not initialized. Retrieving unloaded snapshots.")
+            self.unloaded = True
+            self.update()
         return {k: v for k, v in self.snapshots.items() if not v.loaded}
 
     @property
@@ -166,25 +172,26 @@ class IPFabricAPI(Client):
         else:
             self._snapshot_id = self.snapshots[snapshot_id].snapshot_id
 
-    def get_snapshots(self, loaded_only: bool = True):
-        """
-        Gets all snapshots from IP Fabric and returns a dictionary of {ID: Snapshot_info}
-        :return: dict[str, Snapshot]: Dictionary with ID as key and dictionary with info as the value
-        """
-        payload = {
-            "columns": ['id', 'status', 'finishStatus', 'loadedSize', 'unloadedSize', 'name', 'note', 'sites',
-                        'fromArchive', 'loading', 'locked', 'deviceAddedCount', 'deviceRemovedCount',
-                        'interfaceActiveCount', 'interfaceCount', 'interfaceEdgeCount', 'totalDevCount',
-                        'isLastSnapshot', 'tsChange', 'tsEnd', 'tsStart', 'userCount'],
-            "sort": {"order": "desc", "column": "tsEnd"}
-        }
-        if loaded_only:
-            logger.warning("Retrieving only loaded snapshots. To load all snapshots set X to X.")
-            payload['filters'] = {"and": [{"status": ["eq", "done"]}, {"finishStatus": ["eq", "done"]}]}
-        results = self._ipf_pager('tables/management/snapshots', payload)
-        res = self.get("/snapshots")
-        res.raise_for_status()
-        get_results = {s['id']: s for s in res.json()}
+    def get_snapshot(self, snapshot_id: str):
+        if snapshot_id in self.snapshots:
+            return self.snapshots[snapshot_id]
+        else:
+            payload = {
+                "columns": models.SNAPSHOT_COLUMNS,
+                "filters": {"id": ["eq", snapshot_id]}
+            }
+            results = self._ipf_pager('tables/management/snapshots', payload)
+            if not results:
+                logger.error(f"Snapshot {snapshot_id} not found.")
+                return None
+            get_results = self._get_snapshots()
+            s = results[0]
+            return models.Snapshot(**s, licensedDevCount=get_results[s['id']].get('licensedDevCount', None),
+                                   errors=get_results[s['id']].get('errors', None),
+                                   version=get_results[s['id']]['version'],
+                                   initialVersion=get_results[s['id']].get('initialVersion', None))
+
+    def _get_snapshots(self):
         """
         Need to do a GET and POST to get all Snapshot data. See NIM-7223
         POST Missing:
@@ -193,10 +200,28 @@ class IPFabricAPI(Client):
         version
         initialVersion
         """
+        res = self.get("/snapshots")
+        res.raise_for_status()
+        return {s['id']: s for s in res.json()}
+
+    def get_snapshots(self):
+        """
+        Gets all snapshots from IP Fabric and returns a dictionary of {ID: Snapshot_info}
+        :return: dict[str, Snapshot]: Dictionary with ID as key and dictionary with info as the value
+        """
+        payload = {
+            "columns": models.SNAPSHOT_COLUMNS,
+            "sort": {"order": "desc", "column": "tsEnd"}
+        }
+        if not self.unloaded:
+            logger.warning("Retrieving only loaded snapshots. To load all snapshots set `unloaded` to True.")
+            payload['filters'] = {"and": [{"status": ["eq", "done"]}, {"finishStatus": ["eq", "done"]}]}
+        results = self._ipf_pager('tables/management/snapshots', payload)
+        get_results = self._get_snapshots()
 
         snap_dict = OrderedDict()
         for s in results:
-            snap = models.Snapshot(**s, licensedDevCount=get_results[s['id']]['licensedDevCount'],
+            snap = models.Snapshot(**s, licensedDevCount=get_results[s['id']].get('licensedDevCount', None),
                                    errors=get_results[s['id']].get('errors', None),
                                    version=get_results[s['id']]['version'],
                                    initialVersion=get_results[s['id']].get('initialVersion', None))